Update to v1.1
Combine leader and non-leader log time function
